Bootstrap confidence intervals (#1)
The analyse script now computes confidence intervals to determine
whether enough data has been collected to exclude the possibility
of a timing side channel if the CI is less than 1 cycle, and the
existing statistical tests don't find timing leakage.

The confidence interval is bootstrapped on the pairwise differences
of the worst class pair (i.e. the pair with the lowest p-value),
with a confidence level of 95% by default.

The progress bar has been switched over to the enlighten library
so that multiple progress bars can be shown for the various parallel
computations.